moving display functions to content
