+++ src/components/display/fewlines/IdentitySection/IdentitySection.stories.tsx
@@ -1,11 +1,39 @@
 import React from "react";
 
-import { IdentitySection } from "./IdentitySection";
-import { IdentityTypes } from "@lib/@types";
+import { RightChevron } from "../Icons/RightChevron/RightChevron";
+import { Separator } from "../Separator/Separator";
+import { BoxedLink, IdentitySection } from "./IdentitySection";
+import { Identity, IdentityTypes } from "@lib/@types";
 
 export default {
   title: "components/Identity Section",
   component: IdentitySection,
+};
+
+const displayStandardIdentityList = (
+  identity: Identity,
+  index: number,
+  listLength: number,
+): JSX.Element => {
+  return (
+    <div key={identity.type + index}>
+      <BoxedLink
+        primary={identity.primary}
+        status={identity.status}
+        href={{
+          pathname: "/account/logins/[type]/[id]",
+          query: {
+            type: identity.type.toLowerCase(),
+            id: identity.id,
+          },
+        }}
+      >
+        <p>{identity.value}</p>
+        <RightChevron />
+      </BoxedLink>
+      {index < listLength && <Separator />}
+    </div>
+  );
 };
 
 export const EmailIdentitySection = (): JSX.Element => {
@@ -23,6 +51,7 @@ export const EmailIdentitySection = (): JSX.Element => {
             value: "test@test.test",
           },
         ],
+        displayListMethod: displayStandardIdentityList,
         noIdentityMessage: "No email added yet.",
         addNewIdentityMessage: "add new email address",
       }}
